+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 **Features**
 - Added config option to set the base energy usage of the Controller (default is 0)
 - Added Grid Filter item to filter items in any Grid
+- Added support for processing patterns with big stacksizes
 - Changed Relay recipe to use redstone torch instead of Basic Processor
 
 ### 0.8.12